feat(emacs): save and release existing desktop when saving
If you're using a desktop, if you wanted to stop and use a different
one, you'd have to quit emacs and start it again, then read the
desktop.

I discovered that the desktop-save function has an optional parameter
to release the lock file when saving. Using this, during
my/desktop-read, we can save and release the existing desktop, then
read in another one.

The new function that does this my/save-and-release-desktop, can also
be called interactively.