feat(emacs): add ability to create windows in any direction
I've been looking for a way to be able to either split a window at any
direction or to run a command (i.e. counsel-fzf) and open that in some
kind of split.

I added advice to a few functions, so I could either perform the
command as normal or input a letter to open it in a split instead.

Ivy actions were also added so the results of counsel-fzf, counsel-ag,
etc, could be opened in splits as well.