feat(emacs): use different line numbering dependent on mode
I like to use relative style numbering. In Emacs, this currently poses
one of two issues depending on whether `relative` or `visual` is used.

`relative` is broken when folding code, and `visual` does not work
properly if there are continuation lines (evil mode doesn't count
these lines when using Ex commands like :t or :m).

In other words, the solution is to not fold numbers in relative mode
and to turn off continuation lines/wrapping when in visual mode.

First, the my/get-line-number-type function was added, which selects
which line numbering type to use for the current buffer. If a folding
mode is enabled, or we are using a text mode, we use visual line
numbering. Otherwise, we use relative by default when derived from
prog-mode.

For the other aspect of this, I added code to switch from relative
numbering to visual numbering when enabling a folding mode or
selective display. Additionally, continuation lines/wrapping are
disabled. The inverse is done when leaving or disabling selective
display.