feat(vim): update how sessions work
Previously the session function would save and immediately exit.

I wanted to change this by making the session function only save.

A new variable g:has_session was created to indicate if a session was
in use.

A new autocmd was added to save the session if there is one in use
when exiting.